Refactor writing to a response stream into a new package
