+++ src/test/java/org/veupathdb/service/sr/TestReferences.java
@@ -15,12 +15,12 @@ public class TestReferences extends ReferenceSequencesDAOFactory {
   public static final ReferenceSequenceSpec genomeSpec = new ReferenceSequenceSpec("genome", 10, 10000, true);
   public static final FastaSequenceIndexSqlite genomeIndex = new FastaSequenceIndexSqlite(ref("Genome.fa.fai.sqlite"));
   public static final IndexedFastaSequenceFile genomeSequences = new IndexedFastaSequenceFile(ref("Genome.fa"), genomeIndex);
-  public static final ReferenceSequencesDAO genomeDao = new ReferenceSequencesDAO(genomeSpec, genomeSequences);
+  public static final ReferenceSequencesDAO genomeDao = new ReferenceSequencesDAO(genomeSpec, ref("Genome.fa.fai.sqlite"), ref("Genome.fa"));
 
   public static final ReferenceSequenceSpec proteinSpec = new ReferenceSequenceSpec("protein", 3, 1_000_000, false);
   public static final FastaSequenceIndexSqlite proteinIndex = new FastaSequenceIndexSqlite(ref("AnnotatedProteins.fa.fai.sqlite"));
   public static final IndexedFastaSequenceFile proteinSequences = new IndexedFastaSequenceFile(ref("AnnotatedProteins.fa"), proteinIndex);
-  public static final ReferenceSequencesDAO proteinDao = new ReferenceSequencesDAO(proteinSpec, proteinSequences);
+  public static final ReferenceSequencesDAO proteinDao = new ReferenceSequencesDAO(proteinSpec, ref("AnnotatedProteins.fa.fai.sqlite"), ref("AnnotatedProteins.fa"));
 
   public static void setUp(){
     instances = new HashMap<>();
