+++ src/test/java/org/veupathdb/service/sr/response/BasesTest.java
@@ -1,4 +1,4 @@
-package org.veupathdb.service.sr.util;
+package org.veupathdb.service.sr.response;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.BeforeEach;
